Fix 9 critical+high bugs: YUV→RGB conversion, duplicate route, .toFloat→NaN, sync/complete auth, outsideNow overnight, SSE auth guard
🔴 FaceDetector.kt: imageToBitmap YUV→RGB converter (YOLO now works)
🔴 RegistrationScreen.kt: imageProxyToBitmap YUV→RGB converter
🔴 RegistrationViewModel.kt: pose 'front'→'CENTER' (matching backend)
🔴 events.ts: removed duplicate route + added authGuard to SSE
🔴 metrics.ts: fixed .toFloat NaN bug → proper JS division
🔴 scanMetrics.ts: removed broken include:{student} (no DB relation)
🟡 sync.ts: POST sync/complete validates deviceId from JWT
🟡 attendance.ts/dashboard.ts/report.ts: outsideNow handles overnight

+++ backend/src/routes/metrics.ts
@@ -100,8 +100,8 @@ export const metricRoutes = new Elysia()
         }
       });
 
-      const fpr = reviewedTotal > 0 ? falsePositives.toFloat / reviewedTotal.toFloat : 0;
-      const fnr = reviewedTotal > 0 ? falseNegatives.toFloat / reviewedTotal.toFloat : 0;
+      const fpr = reviewedTotal > 0 ? falsePositives / reviewedTotal : 0;
+      const fnr = reviewedTotal > 0 ? falseNegatives / reviewedTotal : 0;
 
       metrics = {
         id: "",
